fix another conflict

+++ quantization_scripts/quantize_neural_net.py
@@ -32,6 +32,7 @@ fields = [
     'Weight Median', 'Weight Row Max Mean', 
     'Quantization Loss', 'Relative Loss'
 ]
+LAYER_LOGGING_TOGGLE = False
 
 class QuantizeNeuralNet():
     '''
@@ -150,12 +151,12 @@ class QuantizeNeuralNet():
 
         counter = 0
         
-        # logging
-        if os.path.isfile(LAYER_LOG_FILE):
-            os.remove(LAYER_LOG_FILE)
-        with open(LAYER_LOG_FILE, 'w') as f:
-            writer = csv.DictWriter(f, fieldnames=fields)
-            writer.writeheader()
+        if LAYER_LOGGING_TOGGLE:
+            if os.path.isfile(LAYER_LOG_FILE):
+                os.remove(LAYER_LOG_FILE)
+            with open(LAYER_LOG_FILE, 'w') as f:
+                writer = csv.DictWriter(f, fieldnames=fields)
+                writer.writeheader()
         
         for layer_idx in layers_to_quantize:
             gc.collect()
@@ -195,16 +196,11 @@ class QuantizeNeuralNet():
                 # shape (out_channels, in_channesl/groups*k_size[0]*k_size[1])
                 # each row of W is a neuron (vectorized sliding block)
 
-                if groups == 1:
-                    m = analog_layer_input.shape[0]
-                else:
-                    m = analog_layer_input[0].shape[0]
-
                 Q, quantize_error, relative_quantize_error = StepAlgorithm._quantize_layer(W, 
                                             analog_layer_input, 
                                             quantized_layer_input, 
-                                            m,
-                                            self.cnn_alphabet * self.cnn_alphabet_scalar, #* max(groups // 32, 1),
+                                            analog_layer_input.shape[0],
+                                            self.cnn_alphabet * self.cnn_alphabet_scalar,
                                             self.cnn_percentile,
                                             groups=groups
                                             )
@@ -212,10 +208,7 @@ class QuantizeNeuralNet():
                 self.quantized_network_layers[layer_idx].weight.data = torch.Tensor(Q).float().view(W_shape)
             
             print(f'Shape of weight matrix is {W.shape}')
-            if groups == 1:
-                print(f'Shape of X is {analog_layer_input.shape}')
-            else:
-                print(f'Shape of X is {np.vstack(analog_layer_input).shape}')
+            print(f'Shape of X is {analog_layer_input.shape}')
             print(f'Median of W is {np.quantile(np.abs(W), 0.5, axis=1).mean()}')
             print(f'75q of W is {np.quantile(np.abs(W), 0.75, axis=1).mean()}')
             #  print(f'The {round(self.percentile, 2)} percentile of W is {np.quantile(np.abs(W), self.percentile, axis=1).mean()}')
@@ -225,15 +218,15 @@ class QuantizeNeuralNet():
 
             del analog_layer_input, quantized_layer_input
             
-            # logging
-            with open(LAYER_LOG_FILE, 'a') as f:
-                csv_writer = csv.writer(f)
-                row = [
-                    layer_idx, type(self.analog_network_layers[layer_idx]), groups,
-                    np.max(W), np.median(W), np.quantile(np.abs(W), 1, axis=1).mean(),
-                    quantize_error, relative_quantize_error
-                ]
-                csv_writer.writerow(row)
+            if LAYER_LOGGING_TOGGLE:
+                with open(LAYER_LOG_FILE, 'a') as f:
+                    csv_writer = csv.writer(f)
+                    row = [
+                        layer_idx, type(self.analog_network_layers[layer_idx]), groups,
+                        np.max(W), np.median(W), np.quantile(np.abs(W), 1, axis=1).mean(),
+                        quantize_error, relative_quantize_error
+                    ]
+                    csv_writer.writerow(row)
             
             gc.collect()
 
@@ -369,48 +362,74 @@ class SaveInputConv2d:
         # module_in has shape (B, C, H, W)
         module_input = module_in[0]
 
-        if self.groups == 1:
+        # Need to consider both batch_size B and in_channels C
+        unfolded = self.unfolder(module_input)  # shape (B, C*kernel_size[0]*kernel_size[1], L)
+        
+        batch_size, num_blocks = unfolded.shape[0], unfolded.shape[-1]
+        unfolded = torch.transpose(unfolded, 1, 2) # shape (B, L, C*kernel_size[0]*kernel_size[1])
+        unfolded = unfolded.reshape(-1, unfolded.size(-1)).numpy() # shape (B*L, C*kernel_size[0]*kernel_size[1])
 
-            # Need to consider both batch_size B and in_channels C
-            unfolded = self.unfolder(module_input)  # shape (B, C*kernel_size[0]*kernel_size[1], L)
-            
-            batch_size, num_blocks = unfolded.shape[0], unfolded.shape[-1]
-            unfolded = torch.transpose(unfolded, 1, 2) # shape (B, L, C*kernel_size[0]*kernel_size[1])
-            unfolded = unfolded.reshape(-1, unfolded.size(-1)).numpy() # shape (B*L, C*kernel_size[0]*kernel_size[1])
-
-            if self.call_count == 0:
-                self.rand_indices = np.concatenate(
-                            [np.random.choice(np.arange(num_blocks*i, num_blocks*(i+1)), 
-                            size=int(self.p*num_blocks + 1 if self.p != 1 else self.p*num_blocks)) 
-                            for i in range(batch_size)]
-                            ) # need to define self.p (probability)
-            self.call_count += 1
-            unfolded = unfolded[self.rand_indices]
-
-        else:
-            group_channel = module_input.shape[1] // self.groups
-            
-            unfolded = [
-                self.unfolder(module_input[:, i*group_channel: (i+1)*group_channel, :, :])
-                for i in range(self.groups)
-            ]
-
-            batch_size, num_blocks = unfolded[0].shape[0], unfolded[0].shape[-1]
-
-            if self.call_count == 0:
-                self.rand_indices = np.concatenate(
-                            [np.random.choice(np.arange(num_blocks*i, num_blocks*(i+1)), 
-                            size=int(self.p*num_blocks + 1 if self.p != 1 else self.p*num_blocks)) 
-                            for i in range(batch_size)]
-                            ) # need to define self.p (probability)
-            self.call_count += 1
-
-            for i in range(len(unfolded)):
-                unfolded[i] = torch.transpose(unfolded[i], 1, 2)
-                unfolded[i] = unfolded[i].reshape(-1, unfolded[i].size(-1)).numpy()
-                unfolded[i] = unfolded[i][self.rand_indices]
+        if self.call_count == 0:
+            self.rand_indices = np.concatenate(
+                        [np.random.choice(np.arange(num_blocks*i, num_blocks*(i+1)), 
+                        size=int(self.p*num_blocks + 1 if self.p != 1 else self.p*num_blocks)) 
+                        for i in range(batch_size)]
+                        ) # need to define self.p (probability)
+        self.call_count += 1
+        unfolded = unfolded[self.rand_indices]
 
         self.inputs.append(unfolded)
         raise InterruptException
+        
+        
+        # some code might be useful for seperate out channel
+        # if len(module_in) != 1:
+        #     raise TypeError('The number of input layer is not equal to one!')
+        # # module_in has shape (B, C, H, W)
+        # module_input = module_in[0]
+
+        # if self.groups == 1:
+
+        #     # Need to consider both batch_size B and in_channels C
+        #     unfolded = self.unfolder(module_input)  # shape (B, C*kernel_size[0]*kernel_size[1], L)
+            
+        #     batch_size, num_blocks = unfolded.shape[0], unfolded.shape[-1]
+        #     unfolded = torch.transpose(unfolded, 1, 2) # shape (B, L, C*kernel_size[0]*kernel_size[1])
+        #     unfolded = unfolded.reshape(-1, unfolded.size(-1)).numpy() # shape (B*L, C*kernel_size[0]*kernel_size[1])
+
+        #     if self.call_count == 0:
+        #         self.rand_indices = np.concatenate(
+        #                     [np.random.choice(np.arange(num_blocks*i, num_blocks*(i+1)), 
+        #                     size=int(self.p*num_blocks + 1 if self.p != 1 else self.p*num_blocks)) 
+        #                     for i in range(batch_size)]
+        #                     ) # need to define self.p (probability)
+        #     self.call_count += 1
+        #     unfolded = unfolded[self.rand_indices]
+
+        # else:
+        #     group_channel = module_input.shape[1] // self.groups
+            
+        #     unfolded = [
+        #         self.unfolder(module_input[:, i*group_channel: (i+1)*group_channel, :, :])
+        #         for i in range(self.groups)
+        #     ]
+
+        #     batch_size, num_blocks = unfolded[0].shape[0], unfolded[0].shape[-1]
+
+        #     if self.call_count == 0:
+        #         self.rand_indices = np.concatenate(
+        #                     [np.random.choice(np.arange(num_blocks*i, num_blocks*(i+1)), 
+        #                     size=int(self.p*num_blocks + 1 if self.p != 1 else self.p*num_blocks)) 
+        #                     for i in range(batch_size)]
+        #                     ) # need to define self.p (probability)
+        #     self.call_count += 1
+
+        #     for i in range(len(unfolded)):
+        #         unfolded[i] = torch.transpose(unfolded[i], 1, 2)
+        #         unfolded[i] = unfolded[i].reshape(-1, unfolded[i].size(-1)).numpy()
+        #         unfolded[i] = unfolded[i][self.rand_indices]
+
+        # self.inputs.append(unfolded)
+        # raise InterruptException
